adding form functionality to make post requests to db

+++ app/api/matches/[id]/route.ts
@@ -1,5 +1,7 @@
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prismadb";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "../../auth/[...nextauth]/route";
 
 // get match by ID
 export async function GET(
@@ -22,6 +24,14 @@ export async function PUT(
   req: Request,
   { params }: { params: { id: string } }
 ) {
+  //getting the authenticated user session
+  const session = await getServerSession(authOptions);
+
+  //if no session return the next response error
+  if (!session) {
+    return NextResponse.json({ error: "Not Authenticated" }, { status: 401 });
+  }
+
   const { gameMode, kills, deaths, win, time } = await req.json();
   const id = params.id;
   try {
@@ -32,24 +42,31 @@ export async function PUT(
 
     return NextResponse.json(match);
   } catch (error) {
-    return NextResponse.json({message: "Error updating"})
+    return NextResponse.json({ message: "Error updating" });
   }
 }
 
 // delete match by id
 export async function DELETE(
-    req: Request,
-    { params }: { params: { id: string } }
-  ) {
-    const id = params.id;
-    try {
-      const match = await prisma.match.delete({
-        where: { id },
-      });
-  
-      return NextResponse.json(match);
-    } catch (error) {
-      return NextResponse.json({message: "Error deleting"})
-    }
+  req: Request,
+  { params }: { params: { id: string } }
+) {
+  //getting the authenticated user session
+  const session = await getServerSession(authOptions);
+
+  //if no session return the next response error
+  if (!session) {
+    return NextResponse.json({ error: "Not Authenticated" }, { status: 401 });
   }
 
+  const id = params.id;
+  try {
+    const match = await prisma.match.delete({
+      where: { id },
+    });
+
+    return NextResponse.json(match);
+  } catch (error) {
+    return NextResponse.json({ message: "Error deleting" });
+  }
+}
